+++ src/main/java/phanes/replay/gathering/dto/GatheringCreateRequest.java
@@ -1,11 +1,9 @@
-package phanes.replay.gathering.controller;
+package phanes.replay.gathering.dto;
 
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import jakarta.validation.constraints.Future;
 import jakarta.validation.constraints.Min;
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.NotNull;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
